requirements.txt
numpy>=1.21.0
scikit-learn>=0.24.0
pandas>=1.3.0
Pillow>=8.0.0
streamlit>=1.0.0
